chore(kerykeion): split collector tests + repair drifted lint baseline lines
collector.rs crossed the RUST/file-too-long 800-line threshold once
dispatch_routing + its four tests landed; extracted the test module to
collector_tests.rs (same #[path] pattern processor.rs already uses for
processor_tests.rs).

The dispatch_routing insertion also shifted eight pre-existing
RUST/no-arc-mutex-anti-pattern findings in collector.rs and one
RUST/no-silent-result-swallow finding in processor.rs down by a fixed
offset. .kanon-lint-baseline.toml suppression entries are keyed on
(rule, file, line, hash) — same rule/file/hash, corrected line numbers
so these still-accepted pre-existing findings stay suppressed instead
of resurfacing as false new violations from an unrelated insertion.

Gate-Passed: kanon 0.1.10 +stages:fmt,check,clippy,nextest,lint sha:95285a93109a3d91b8c729c566b6680699f7b2ef
